uc/cucm-new-e164-patterns/requirements.txt
appdirs==1.4.3
attrs==19.1.0
cached-property==1.5.1
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
idna==2.8
isodate==0.6.0
lxml==4.4.1
pyaml==19.4.1
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
requests-toolbelt==0.9.1
six==1.12.0
urllib3==1.25.3
zeep==3.4.0
